refactoring for circular import
